+++ skilfor/src/components/Calendar/CourseCalendar.js
@@ -4,92 +4,11 @@ import { Calendar, momentLocalizer } from "react-big-calendar";
 import moment from "moment";
 import "react-big-calendar/lib/css/react-big-calendar.css";
 import events from "./event";
+import AddTaskAlertCard from "./AddTaskAlertCard";
 
 const CalendarContainer = styled.div`
   position: relative;
 `;
-
-const RowContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: end;
-`;
-
-const AddNewContainer = styled.div`
-  position: absolute;
-  left: 50%;
-  top: 50%;
-  transform: translate(-50%, -50%);
-  width: 300px;
-  background: white;
-  border-top: 8px solid ${(props) => props.theme.colors.green_dark};
-  padding: 20px;
-  text-align: center;
-  box-shadow: 0 10px 20px 10px rgba(0, 0, 0, 0.2);
-  z-index: 5;
-`;
-
-const AddNewTitle = styled.h1`
-  font-size: 1.2rem;
-  color: ${(props) => props.theme.colors.grey_dark};
-  margin: 0 auto;
-  width: fit-content;
-`;
-
-const AddNewContent = styled.p`
-  font-size: 1.1rem;
-  margin-top: 10px;
-  color: ${(props) => props.theme.colors.grey_dark};
-`;
-
-const SelectContainer = styled.select`
-  height: 25px;
-  font-size: 1rem;
-`;
-
-const SelectOption = styled.option`
-  over-flow: scroll;
-`;
-
-const AddButton = styled.button`
-  border: none;
-  border-radius: 5px;
-  width: 80px;
-  color: white;
-  background: ${(props) => props.theme.colors.green_dark};
-  margin: 0 auto;
-  margin-top: 15px;
-  padding: 6px 10px;
-  cursor: pointer;
-  :hover {
-    opacity: 0.8;
-  }
-`;
-
-const createTimeOptionsList = () => {
-  let arr = [];
-  for (let i = 0; i < 24; i++) {
-    arr.push(`${i}:00`);
-    arr.push(`${i}:30`);
-  }
-  arr.push("24:00");
-  return arr;
-};
-const createTimeOptions = (timeType, time) => {
-  let list = createTimeOptionsList();
-  //沒給定開始時間的結束時間選單
-  if (!time && timeType === "end") {
-    return list.slice(1, list.length);
-  }
-  //給定開始時間的結束時間選單
-  if (time && timeType === "end") {
-    return list.slice(list.indexOf(time) + 1, list.length);
-  }
-  //開始時間選單
-  if (timeType === "start") {
-    return list.slice(0, list.length - 1);
-  }
-};
 
 function CourseCalendar({ courseName }) {
   const localizer = momentLocalizer(moment);
@@ -98,14 +17,6 @@ function CourseCalendar({ courseName }) {
     title: courseName,
     start: "",
     end: "",
-  });
-  const [courseTime, setCourseTime] = useState({
-    start: {
-      time: "0:00",
-    },
-    end: {
-      time: "0:30",
-    },
   });
 
   const handleDateClick = (e) => {
@@ -116,47 +27,8 @@ function CourseCalendar({ courseName }) {
     console.log(e);
   };
 
-  const handleCourseTimeChange = (e) => {
-    const { id, value } = e.target;
-    setCourseTime({
-      ...courseTime,
-      [id]: {
-        time: value,
-      },
-    });
-  };
-
   const MyCalendar = (props) => (
     <>
-      <AddNewContainer>
-        <AddNewTitle>新增一個上課時段</AddNewTitle>
-        <AddNewTitle>11月17日(星期三)</AddNewTitle>
-        <RowContainer>
-          <AddNewContent>開始時間：</AddNewContent>
-          <SelectContainer
-            id="start"
-            onChange={handleCourseTimeChange}
-            value={courseTime.start.time}
-          >
-            {createTimeOptions("start", courseTime.end.time).map((item) => (
-              <SelectOption>{item}</SelectOption>
-            ))}
-          </SelectContainer>
-        </RowContainer>
-        <RowContainer>
-          <AddNewContent>結束時間：</AddNewContent>
-          <SelectContainer
-            id="end"
-            onChange={handleCourseTimeChange}
-            value={courseTime.end.time}
-          >
-            {createTimeOptions("end", courseTime.start.time).map((item) => (
-              <SelectOption>{item}</SelectOption>
-            ))}
-          </SelectContainer>
-        </RowContainer>
-        <AddButton>確定新增</AddButton>
-      </AddNewContainer>
       <Calendar
         onSelectEvent={handleEventClick}
         onSelectSlot={handleDateClick}
@@ -171,6 +43,7 @@ function CourseCalendar({ courseName }) {
   );
   return (
     <CalendarContainer>
+      <AddTaskAlertCard />
       <MyCalendar />
     </CalendarContainer>
   );
